requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.3.2
distro==1.9.0
Django==5.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
jiter==0.6.1
newsapi-python==0.2.7
openai==1.51.1
packaging==24.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
redis==5.1.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
spotipy==2.24.0
sqlparse==0.5.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
